Create allocator.h file
